barber-nyumbani-api/barber-nyumbani-c6a825f4e67ed23b6f61bd3efd90de1c1e7a87c6/requirements.txt
africastalking==1.2.4
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
cloudinary==1.26.0
confusable-homoglyphs==3.2.0
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.7
django-bootstrap-v5==1.0.5
django-cors-headers==3.10.0
django-heroku==0.3.1
django-mathfilters==1.0.0
django-registration==3.2
django-tinymce==3.3.0
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==3.10.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.2
Pillow==8.3.2
psycopg2==2.9.1
pycodestyle==2.7.0
pyparsing==2.4.7
python-decouple==3.5
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
schema==0.7.4
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
